+++ src/main/java/org/itxtech/synapseapi/multiprotocol/protocol116220/protocol/CraftingDataPacket116220.java
@@ -19,7 +19,7 @@ import java.util.stream.Collectors;
 @ToString
 public class CraftingDataPacket116220 extends Packet116220 {
 
-    public static final byte NETWORK_ID = ProtocolInfo.CRAFTING_DATA_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.CRAFTING_DATA_PACKET;
 
     public static final String CRAFTING_TAG_CRAFTING_TABLE = "crafting_table";
     public static final String CRAFTING_TAG_CARTOGRAPHY_TABLE = "cartography_table";
