+++ src/main/java/org/itxtech/synapseapi/multiprotocol/protocol11730/protocol/CraftingDataPacket11730.java
@@ -20,7 +20,7 @@ import java.util.stream.Collectors;
 @ToString
 public class CraftingDataPacket11730 extends Packet11730 {
 
-    public static final byte NETWORK_ID = ProtocolInfo.CRAFTING_DATA_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.CRAFTING_DATA_PACKET;
 
     public static final String CRAFTING_TAG_CRAFTING_TABLE = "crafting_table";
     public static final String CRAFTING_TAG_CARTOGRAPHY_TABLE = "cartography_table";
