+++ src/main/java/org/itxtech/synapseapi/multiprotocol/protocol12010/protocol/AvailableCommandsPacket12010.java
@@ -23,7 +23,7 @@ import java.util.function.ObjIntConsumer;
 
 @ToString
 public class AvailableCommandsPacket12010 extends Packet12010 {
-    public static final byte NETWORK_ID = ProtocolInfo.AVAILABLE_COMMANDS_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.AVAILABLE_COMMANDS_PACKET;
 
     private static final ObjIntConsumer<BinaryStream> WRITE_BYTE = (s, v) -> s.putByte((byte) v);
     private static final ObjIntConsumer<BinaryStream> WRITE_SHORT = BinaryStream::putLShort;
