+++ src/main/java/org/itxtech/synapseapi/multiprotocol/protocol15/protocol/SetLocalPlayerAsInitializedPacket15.java
@@ -5,7 +5,7 @@ import lombok.ToString;
 
 @ToString
 public class SetLocalPlayerAsInitializedPacket15 extends Packet15 {
-    public static final byte NETWORK_ID = ProtocolInfo.SET_LOCAL_PLAYER_AS_INITIALIZED_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.SET_LOCAL_PLAYER_AS_INITIALIZED_PACKET;
 
     public long eid;
