+++ src/main/java/org/itxtech/synapseapi/multiprotocol/protocol16/protocol/SetLocalPlayerAsInitializedPacket16.java
@@ -5,7 +5,7 @@ import lombok.ToString;
 
 @ToString
 public class SetLocalPlayerAsInitializedPacket16 extends Packet16 {
-    public static final byte NETWORK_ID = ProtocolInfo.SET_LOCAL_PLAYER_AS_INITIALIZED_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.SET_LOCAL_PLAYER_AS_INITIALIZED_PACKET;
 
     public long eid;
