+++ src/main/java/org/itxtech/synapseapi/multiprotocol/protocol18/protocol/AvailableEntityIdentifiersPacket18.java
@@ -3,7 +3,7 @@ package org.itxtech.synapseapi.multiprotocol.protocol18.protocol;
 import cn.nukkit.network.protocol.ProtocolInfo;
 
 public class AvailableEntityIdentifiersPacket18 extends Packet18 {
-    public static final byte NETWORK_ID = ProtocolInfo.AVAILABLE_ACTOR_IDENTIFIERS_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.AVAILABLE_ACTOR_IDENTIFIERS_PACKET;
 
     /**
      * CompoundTag.
